Some refactoring work made to DiskManager to make it simpler (shorter methods) and more readable.
PieceMapEntry moved out from DiskManager.

PieceList created instead of the List used previously.

Piece has a PeerManager which is only used in the UI in one location...
note made that this should be removed.

In the future, it might be desirable to move the Piece*.java classes into
their own package (core.pieces)

-Moti


git-svn-id: http://svn.vuze.com/public/client/trunk@463 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2