In the process method, we adjust the _connection_id attribute of any RPPluginInterface object we are working with to match the one we have on the request.
We do this, because newly created RPPluginInterface objects increment the connection ID to use. This is normally fine, but recent changes have meant that plugin interfaces can actually get references to *other* plugin interfaces, so we have to keep the connection ID's consistent with what is in the request (otherwise, caller code may believe that the connection ID has changed).


git-svn-id: http://svn.vuze.com/public/client/trunk@8360 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2