Note to other devs: Don't listen to users. ;)
The "Restart Azureus Now" / "Restart Azureus Later" text is just too long. If we really have to make it clear whether we are referring to the application or the operating system, it should be made clear as part of the text, rather than the buttons.

I was going to revert the language text for other languages - but I'm not convinced I can do it properly.


git-svn-id: http://svn.vuze.com/public/client/trunk@12341 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2