We now have two different attribute listening frameworks in place now - the existing DMSListener and the new DMSAttributeListener framework. The older one is being deprecated in favour of the new one.
The DMSListener class ended up firing off a lot of events, because it told listeners about both read and write events for all attributes. To avoid this, the new interface requires you to explicitly mention which attribute and what event type (to-be-read or written).

So - a tiny bit more awkward to register and deregister, but should result in a significant decrease the number of events we generate (as well as the number of event objects that the old API uses).


git-svn-id: http://svn.vuze.com/public/client/trunk@13125 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2