Fairly big change to the default save location manager code - the functionality is now split into two separate parts.
DownloadManagerMoveHandler is the code which is responsible for deciding where download should go to (based on what save location manager is use). It's the public API to interact with it.

DownloadManagerDefaultPaths now contains the manager object which embodies our move-on-completion and move-on-removal logic - which is presented in a save location manager form - you shouldn't really need to interact with this class.


git-svn-id: http://svn.vuze.com/public/client/trunk@13252 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2