The virtual table behaviour we present for plugins (to get it working in the torrent menu) was using the wrong table for incomplete downloads - now fixed.
On a side note, please don't change other people's code just to reformat it - it's very annoying. Unless you're changing that particular code block - don't do it.

We all have our own preferences over code formatting - but don't change code just for the sake of it. It screws up the CVS blame history for a start...


git-svn-id: http://svn.vuze.com/public/client/trunk@13606 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2