The getPluginInterfaceByXXX methods now have a variant which allow you to indicate whether you care if the plugin is operational or not.
By default, the existing methods will now only return true if the plugin is operational - that deviates from the old behaviour. I doubt this is likely to introduce bugs (I reckon that it's more likely to fix existing bugs), but if it is causing problems, the behaviour can be switched back by modifying the constant defined in PluginManagerImpl.


git-svn-id: http://svn.vuze.com/public/client/trunk@13894 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2